+++ snippets/usage.rs
@@ -1,12 +1,18 @@
 use compose_idents::compose_idents;
 
 compose_idents!(
+    // Valid identifiers, underscores, integers and strings are allowed as literal values.
     my_fn_1 = [foo, _, "baz"];
     my_fn_2 = [spam, _, 1, _, eggs];
+    // Functions can be applied to the arguments.
     my_const = [upper(foo), _, lower(BAR)];
+    // Function calls can be arbitrarily nested and combined.
     my_static = [upper(lower(BAR))];
     MY_SNAKE_CASE_STATIC = [snake_case(snakeCase)];
     MY_CAMEL_CASE_STATIC = [camel_case(camel_case)];
+    // This function is useful to create identifiers that are unique across multiple macro invocations.
+    // `hash(0b11001010010111)` will generate the same value even if called twice in the same macro call,
+    // but will be different in different macro calls.
     MY_UNIQUE_STATIC = [hash(0b11001010010111)]; {
     fn my_fn_1() -> u32 {
         123
@@ -23,6 +29,7 @@ compose_idents!(
     static MY_UNIQUE_STATIC: u32 = 42;
 });
 
+// It's possible to use arguments of declarative macros as parts of the identifiers.
 macro_rules! outer_macro {
     ($name:tt) => {
         compose_idents!(my_nested_fn = [nested, _, $name]; {