+++ tests/tests.rs
@@ -2,6 +2,7 @@
 fn compile_tests() {
     let t = trybuild::TestCases::new();
     t.pass("tests/compile/simple_compose.rs");
+    t.pass("tests/compile/trailing_semicolon.rs");
     t.pass("tests/compile/nested_compose.rs");
     t.pass("tests/compile/nested_type_token_compose.rs");
     t.pass("tests/compile/multi_compose.rs");