+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Documented functions and add more clarity to the docs in general.
+- Updated syn version to 2.
 
 ### Changed